员工不能更改名称

+++ app/src/main/java/com/heloo/android/osmapp/model/OrderPriceBO.java
@@ -16,6 +16,6 @@ public class OrderPriceBO {
         public String icon;
         public Integer number;
         public Double discountprice;
-        public Integer integralprice;
+        public Double integralprice;
     }
 }
